==> 2_full_data_individual_samples/1a_run_vsn_atac_qc.ipynb <==
nwork=${VSC_SCRATCH}/HCA_PBMC_FULL_DATA_QC/
mkdir $nwork
export NXF_WORK=$nwork

module load Nextflow
module load graphviz




nextflow pull vib-singlecell-nf/vsn-pipelines -r develop_atac

VSN=vib-singlecell-nf/vsn-pipelines/main_atac.nf

Then, generate a vsn config for the atac_qc pipeline. Put the right parameters:
* change fragments input files glob e.g. `fragments_bap/fragments_mm/*.tsv.gz`
* change suffix to `.sinto.fragments.tsv.gz` or make sure it matches the files. only sample name should remain. make sure this is correct
* change fragments input files glob e.g. `fragments_bap/bam_postbap/*.bam`
* change suffix to `.bwa.out.possorted.bam` or make sure it matches the files. only sample name should remain.
* edit cell calling, can be sample-specific
* edit `out` to `out_qc`

nextflow config $VSN \
    -profile atac_qc_filtering,bam,fragments,pycistopic_hg38,vsc \
    > atac_qc.config

Make sure to change all the parameters:

Then, in a tmux session:

nextflow -C atac_qc_all.config run \
    vib-singlecell-nf/vsn-pipelines/main_atac.nf \
    -entry atac_qc_filtering -r develop_atac -resume --quiet

After this is finished, read the bap reports to check the bap kneeplot and determine new bap parameters. Re-run the pipeline with new bap parameters.
