==> 2_full_data_individual_samples/2b_fric.ipynb <==
import pandas as pd
import glob
import os
from collections import OrderedDict

filenames = glob.glob('selected_barcodes/*')

filenames = sorted(glob.glob('selected_barcodes/*'))
samples = [item.replace("selected_barcodes/", "") for item in filenames]
samples = [item.replace(".cell_barcodes.txt", "") for item in samples]
bc_dict = {samples[i]: filenames[i] for i in range(len(samples))}
bc_dict

filenames = sorted(glob.glob('fragments_postbap/*tsv'))
samples = [item.replace("fragments_postbap/", "") for item in filenames]
samples = [item.replace(".sinto.mm.fragments.tsv", "") for item in samples]
frag_dict = {samples[i]: filenames[i] for i in range(len(samples))}
frag_dict

for key in bc_dict.keys():
    print(key)
    df_frags = pd.read_csv(frag_dict[key], sep='\t', header=None, index_col=None)
    df_frags.columns = "chr start end bc count".split()
    df_bc = pd.read_csv(bc_dict[key], header=None, index_col=None)
    df_bc.columns = "bc".split()


    df_frags_incells = df_frags[df_frags['bc'].isin(df_bc['bc'])]
    frac_reads = df_frags_incells['count'].sum()/df_frags['count'].sum()
    frac_fragments = len(df_frags_incells)/len(df_frags)

    print(f'\ttotal reads: {df_frags["count"].sum()}')
    print(f'\tcells pre-scrublet/freemuxlet: {len(df_bc)}')
    print(f'\tmapped reads per cell: {df_frags["count"].sum()/len(df_bc)}')
    print(f'\tfraction mapped fragments in cells: {frac_reads}')
    print(f'\tfraction unique fragments in cells: {frac_fragments}')

for key in bc_dict.keys():
    print(key)
    df_frags = pd.read_csv(frag_dict[key], sep='\t', header=None, index_col=None)
    df_frags.columns = "chr start end bc count".split()
    df_bc = pd.read_csv(bc_dict[key], header=None, index_col=None)
    df_bc.columns = "bc".split()


    df_frags_incells = df_frags[df_frags['bc'].isin(df_bc['bc'])]
    frac_reads = df_frags_incells['count'].sum()/df_frags['count'].sum()
    frac_fragments = len(df_frags_incells)/len(df_frags)

    print(f'\ttotal reads: {df_frags["count"].sum()}')
    print(f'\tcells pre-scrublet/freemuxlet: {len(df_bc)}')
    print(f'\tmapped reads per cell: {df_frags["count"].sum()/len(df_bc)}')
    print(f'\tfraction mapped fragments in cells: {frac_reads}')
    print(f'\tfraction unique fragments in cells: {frac_fragments}')

filenames = sorted(glob.glob('mapping_stats/barcode/*'))
samples = [item.replace("/lustre1/project/stg_00002/lcb/cflerin/analysis/pbmc_atac/analysis2/data_freeze_Jun2021_library_downsampled/atac_preprocess/out/data/reports/barcode/", "") for item in filenames]
samples = [item.replace(".mapping_stats.tsv", "") for item in samples]
stats_dict = {samples[i]: filenames[i] for i in range(len(samples))}
stats_dict

key = 'VIB_1'
