==> 2_full_data_individual_samples/4b_export_looms_celltype_pred.ipynb <==
# pycisTopic analysis

Full dataset, using SCREEN regions.

import pycisTopic
pycisTopic.__version__

import warnings
warnings.filterwarnings('ignore')
warnings.simplefilter('ignore')

import pickle
import pandas as pd

import os
wdir = '/lustre1/project/stg_00002/lcb/fderop/data/20211008_hca_benchmark_full_singles/'
os.chdir( wdir )

import glob
from collections import OrderedDict

# load cistopic objs

filenames = glob.glob('fragments_postbap/*.sinto.mm.fragments.tsv.gz')
samples = [item.replace(".sinto.mm.fragments.tsv.gz", "") for item in filenames]
samples = [item.replace("fragments_postbap/", "") for item in samples]
files_dict = {samples[i]: filenames[i] for i in range(len(samples))}
files_dict = OrderedDict(sorted(files_dict.items()))
files_dict.keys()

f_cto_dir = 'cistopic_objs__screen'

cistopic_obj_dict = {}
for key in files_dict.keys():
    f_cto = os.path.join(wdir, f_cto_dir, key + '__cistopic_obj_filtered.pkl')
    if(os.path.isfile(f_cto)):
        with open(f_cto, 'rb') as f:
            cistopic_obj_dict[key] = pickle.load(f)
        print(f"Loaded filtered cistopic object {key}")
    else:
        print(f"file {f_cto} doesn't exist")

## Export region-accessibility looms

For use in cell type identification
# BE CAREFUL WITH THIS: col_attrs = barcodes -> need to make sure that the barcode will match the one in the fragments files, otherwise downstream will not work.

better is for example: `            col_attrs={ 'CellID': list(cistopic_obj_dict[key].cell_data['barcode']) }, # for hydrop
`

import loompy as lp

f_crl_dir = 'cell_region_loom__screen'
if not os.path.exists(os.path.join(wdir, f_crl_dir)):
    os.makedirs(os.path.join(wdir, f_crl_dir))

for key in cistopic_obj_dict.keys():
    f_out = os.path.join(wdir, f_crl_dir, key + 'cell_region-all.loom')
    if os.path.isfile(f_out):
        print(f"Skipping {key} loom writing, already exists")
    else:
        lp.create(
            filename = f_out,
            layers=cistopic_obj_dict[key].fragment_matrix,
            row_attrs={ 'Gene': cistopic_obj_dict[key].region_names }, 
            #col_attrs={ 'CellID': cistopic_obj_dict[key].cell_names }, 
            # col_attrs={ 'CellID': [ x.split('-')[0] for x in cistopic_obj_dict[key].cell_names ] },
            col_attrs={ 'CellID': [ x.split('-')[0] + '-' + x.split('-')[1]  for x in cistopic_obj_dict[key].cell_names ] },

            #file_attrs=attrs
        )
        print(f"Finished {key} loom writing")

go to next notebook to run the modeling

cistopic_obj_dict['VIB_Hydrop_1'].cell_data

[ x.split('-')[0] for x in cistopic_obj_dict['VIB_Hydrop_1'].cell_names ]
